refactor(data_loader): Add data loading and processing utilities for robot calibration
